Support tie-break with arbitrary integer points
With this change, votes are not limited to Eurovision-style points only.
This is useful for competitions that, for example, extend the point
scale to include numbers 9, 11, or larger numbers than 12. This patch
supports integer values only.

This approach finds all unique points and caches them in order to be
used in result sorting. In fact, the points could be counted in the same
pass but the counts would have to be stored for each voter. Thus, only
the unique points are stored.

Another approach could be to require the user to configure valid point
values beforehand but it would require an additional configuration step.
On the other hand, the known values could be used for validation as well
in order to check missing points.

The change is made in the final_results() method as well although it
seems like it is not called anywhere.