Remember previously selected folder during session
This change attempts to slightly improve usability by remembering the
previously selected directory when choosing the input file or output
folder. The selected location is shared between input file location and
output folder, that is, the directory changes when either of them is
changed, The selection is not remembered between sessions, though.

Before this change, one had to navigate to the desired output location
starting from the home directory. If the user wants to store the output
pictures close to the input file, this navigation seems unnecessary. In
this case, remembering the previous location provides a quick way to get
to near the desired location, which was the motivation for this patch.

Another approach could be to remember input and output folders
separately. This might be beneficial if the user wants to store output
in completely different location than the input file.

Remembering the location also helps in case the user wants to reselect
a location for some reason, for example, due to a misclick.